# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple Swagger implementations in GO HTTP",
        "title": "Albums RESTful API with Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Roberto Guzmán",
            "email": "roberto140298@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/albums": {
            "get": {
                "description": "gets all albums in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "List albums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Album"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing album in the database corresponding to the id sent. Otherwise, returns error",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Update an album",
                "parameters": [
                    {
                        "description": "Album JSON Object",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Create a new album",
                "parameters": [
                    {
                        "description": "Album JSON Object",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/albums/{id}": {
            "get": {
                "description": "Gets a single album from the database corresponding to the id in the path. Otherwise, returns error",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Get an album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Album": {
            "description": "album information with album id, album title, arist and price",
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}